Completed Spring Data JPA Quick Example with Criteria API

+++ StudentManagementApplication/management/src/main/java/com/digitalnurture/student/entity/Course.java
@@ -1,15 +1,15 @@
 package com.digitalnurture.student.entity;
 
+import com.fasterxml.jackson.annotation.JsonBackReference;
 import jakarta.persistence.*;
 import java.util.List;
-import com.fasterxml.jackson.annotation.JsonBackReference;
 
 @Entity
 public class Course {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String title;
 
@@ -19,18 +19,15 @@ public class Course {
 
     public Course() {}
 
-    public Course(Integer id, String title) {
+    public Course(Long id, String title) {
         this.id = id;
         this.title = title;
     }
 
- 
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
-
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
     public String getTitle() { return title; }
     public void setTitle(String title) { this.title = title; }
-
     public List<Student> getStudents() { return students; }
     public void setStudents(List<Student> students) { this.students = students; }
 }
